refactor(runtime): isolate wrapper host authentication emission
